+++ lib/src/domain/user_options/models/invalid_user_option_error.dart
@@ -1,0 +1,9 @@
+class InvalidUserOptionError implements Exception {
+  final String optionName;
+  final String source;
+
+  const InvalidUserOptionError(this.optionName, this.source);
+
+  @override
+  String toString() => "Invalid option `$optionName` on $source options";
+}