+++ lib/src/domain/user_options/use_case/get_or_fail_user_options.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:pull_request_coverage/src/domain/analyzer/models/exit_code.dart';
 import 'package:pull_request_coverage/src/domain/common/result.dart';
+import 'package:pull_request_coverage/src/domain/user_options/models/invalid_user_option_error.dart';
 import 'package:pull_request_coverage/src/domain/user_options/models/user_options.dart';
 import 'package:pull_request_coverage/src/domain/user_options/repositories/user_options_repository.dart';
 
@@ -16,7 +17,12 @@ class GetOrFailUserOptions {
       return userOptions.data;
     } else {
       userOptions as ResultError<UserOptions>;
-      print("Error parsing params: ${userOptions.message}\n${userOptions.stackTrace}");
+      final error = userOptions.error;
+      if (error is InvalidUserOptionError) {
+        print(error.toString());
+      } else {
+        print("Error parsing params: ${userOptions.message}\n${userOptions.stackTrace}");
+      }
       exit(ExitCode.error);
     }
   }