+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Output now has emojis by default. You can disable it by setting `print-emojis` to `false`.
 - Output is now sorted: files missing tests will be shown at the end of the report
 - Change output style
+- Handling invalid options
 
 ## 1.3.6